Move isAllASCII from UTF8Helpers to StringUtils
