+++ src/Common/UTF8Helpers.h
@@ -136,10 +136,6 @@ size_t computeWidth(const UInt8 * data, size_t size, size_t prefix = 0) noexcept
   */
 size_t computeBytesBeforeWidth(const UInt8 * data, size_t size, size_t prefix, size_t limit) noexcept;
 
-
-/// If all the characters in the string are ASCII, return true.
-bool isAllASCII(const UInt8* data, size_t size);
-
 }
 
 }
